test: restore integration tests
