+++ tests/Cargo.toml
@@ -12,11 +12,13 @@ path = "integrations/lib.rs"
 anyhow = "1.0"
 bytes = "1"
 bytesize = "1.1.0"
+clap = { version = "3.1.6", features = ["derive"] }
 env_logger = "*"
 futures = "0.3"
 itertools = "0.10.3"
 lazy_static = "1.4.0"
 rand = "0.8.5"
+runkv-client = { path = "../client" }
 runkv-common = { path = "../common" }
 runkv-exhauster = { path = "../exhauster" }
 runkv-proto = { path = "../proto" }
